backup_model/v1/requirements-api.txt
flask>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0
cachetools>=5.3.0
pyzmq>=26.0.0
msgpack>=1.0.8
